Add title here.

Add more detailed description here. Explain that this solves the dual problem formulation using the MOSEK interior-point solver (see text from manuscript).

mixKWDual(L, w = rep(1, nrow(L)), ...)

Arguments

L
Matrix specifying the optimization problem to be solved. It should be a numeric matrix with positive entries, and ideally double-precision.

w
A numeric vector, with one entry for each row of L, specifying the "weights" associated with the rows of L. All weights must be positive. It is assumed the weights sum to 1; if not, they will automatically be normalized to sum to 1. By default, all rows of L are assigned the same weight.

...
Additional optimization parameters passed to MOSEK. See KWDual for details.

Value

mixKWDual returns a list with two components:

x

The solution to the optimization problem, as provided by MOSEK.

value

The value of the objective at x.

status

The MOSEK convergence status.

Examples

# Add example here.