mixSQP

mixSQP solves a convex optimization problem in https://arxiv.org/abs/1806.01412. It implements a sequential quadratic programming with active-set subproblem solvers.

When L is a (n) by (m) matrix of nonnegative entries, mixSQP maximizes the following objective function $$f(x) = \sum_j w_j \log (\sum_k L_jk x_k)$$ subject to the (unit) probability simplex constraint $$\sum_k x_k = 1, x_k \ge 0$$ Without loss of generality \(\sum_j w_j = 1\) is required.

mixKWDual(L, w = rep(1, nrow(L)), ...)

mixobjective(L, x, w = rep(1, nrow(L)))

mixSQP(L, w = rep(1, nrow(L)), x0 = rep(1, ncol(L)),
  convtol.sqp = 1e-08, zero.threshold = 1e-06,
  eps = .Machine$double.eps, delta = 0.01 * convtol.sqp,
  maxiter.sqp = 1000, maxiter.activeset = 100, verbose = TRUE)

Arguments

L
Matrix specifying the optimization problem to be solved. It should be a numeric matrix with at least two columns, with finite and positive entries, and, ideally, stored in double-precision.

w
A numeric vector, with one entry for each row of L, specifying the "weights" associated with the rows of L. All weights must be finite and positive. The weights should sum to 1; if not, they will automatically be normalized to sum to 1. By default, all weights are set to 1, in which case all rows of L are assigned the same weight.

...
Additional arguments passed to KWDual.

x0
An initial estimate of the solution to the optimization problem. It should contain only finite, non-negative values, and the entries should sum to 1; if not, the vector is automatically normalized to sum to 1. By default, x0 is the vector with all equal entries.

convtol.sqp
A convergence tolerance used for algorithm's convergence criterion.

zero.threshold
A tolerance used for determining active indices.

eps
A small constant to safeguard from a numerical issue.

delta
A small constant added to diagonal of Hessian to improve numerical stability of search directions in active-set method.

maxiter.sqp
Maximum number of SQP iterations; that is, the maximum number of quadratic subproblems that will be solved by the active-set method.

maxiter.activeset
Maximum number of active-set iterations taken to solve each of the quadratic subproblems.

verbose
If verbose = TRUE, print progress of algorithm to console.

Value

Returns a solution x (in the current version). Also returns the algorithm convergence status.

Examples

set.seed(1)
n  <- 1e5
m  <- 10
w  <- rep(1,n)/n
L  <- simulatemixdata(n,m)$L
out.mixsqp <- mixSQP(L,w)

#> Running mix-SQP 0.1-22 on 100000 x 10 matrix
#> convergence tol. (SQP): 1.0e-08
#> max. iter (SQP):        1000
#> max. iter (active-set): 100
#> zero threshold:         1.0e-06
#> iter  objective   -min(g+1) #nnz #nqp #nls
#>    1  7.72664e-01 +2.76e-01   10 
#>    2  6.53047e-01 +4.75e+00    5   13    1
#>    3  6.04839e-01 +2.16e+00    5    8    1
#>    4  5.99876e-01 +9.17e-01    5    2    1
#>    5  5.99756e-01 +2.98e-01    5    2    1
#>    6  5.99697e-01 +1.65e-01    5    2    1
#>    7  5.99516e-01 +7.18e-02    5    2    1
#>    8  5.99192e-01 +1.96e-02    5    2    1
#>    9  5.98988e-01 +1.97e-03    5    2    1
#>   10  5.98959e-01 +2.23e-05    5    2    1
#>   11  5.98959e-01 +2.89e-09    5    2    1
#> Convergence criteria met---optimal solution found.

out.kwdual <- mixKWDual(L,w)
print(mixobjective(L,out.mixsqp$x,w),digits = 16)

#> [1] 0.5989590363728408

print(mixobjective(L,out.kwdual$x,w),digits = 16)

#> [1] 0.5989590363746571