Maximum-likelihood estimation of mixture proportions using MOSEK

This function is simply a wrapper to KWDual with a similar interface to mixsqp for solving the same problem as mixsqp. See mixsqp and KWDual for details.

mixkwdual(L, w = rep(1, nrow(L)), ...)

Arguments

L
Matrix specifying the optimization problem to be solved. In the context of mixture-model fitting, L[j,k] should be the value of the kth mixture component density at the jth data point. L should be a numeric matrix with at least two columns, with all entries being non-negative and finite (and not missing). For large matrices, it is preferrable that the matrix is stored in double-precision; see storage.mode.

w
An optional numeric vector, with one entry for each row of L, specifying the "weights" associated with the rows of L. All weights must be finite, non-negative and not missing. Internally, the weights are normalized to sum to 1, which does not change the problem, but does change the value of the objective function reported. By default, all weights are equal.

...
Additional arguments passed to KWDual.

Value

A list object with the following elements:

x

The estimated solution to the convex optimization problem.

value

The value of the objective function at x.

status

The return status from MOSEK.

For more information on this output, see KWDual and mosek.
See also