Solution to "Mixture Optimization" Problem

mixSQP and mixKWDual can be used to compute maximum-likelihood estimates of mixture proportions in a (finite) mixture model, or it can be used more generally to solve the a constrained, convex optimization problem of the form given below (see "Details"). mixSQP uses a Sequential Quadatric Programming (SQP) approach to solve a slightly modified primal formulation of the convex optimization problem. See "References" for more details about the SQP algorithm and the motivation behind it. mixKWDual uses the MOSEK interior-point (IP) algorithm to solve a dual formulation of the original problem; see KWDual for details.

mixSQP(L, w = rep(1, nrow(L)), x0 = rep(1, ncol(L)),
  convtol.sqp = 1e-08, convtol.activeset = 1e-10,
  zero.threshold = 1e-06, eps = .Machine$double.eps, delta = 1e-10,
  maxiter.sqp = 1000, maxiter.activeset = 100, verbose = TRUE)

mixobjective(L, x, w = rep(1, nrow(L)))

mixKWDual(L, w = rep(1, nrow(L)), ...)

Arguments

L
Matrix specifying the optimization problem to be solved. In the context of mixture-model fitting, L[j,k] should be the response of the kth mixture component density at the jth data point. L should be a numeric matrix with at least two columns, in which all its entries are positive finite (and not missing). For large matrices, it is preferrable that the matrix is stored in double-precision; see storage.mode.

w
An optional numeric vector, with one entry for each row of L, specifying the "weights" associated with the rows of L. All weights must be finite, non-negative and not missing. The weights should sum to 1; if they are not, they will automatically be normalized to sum to 1. By default, all weights are the same.

x0
An optional numeric vector providing an initial estimate of the solution to the optimization problem. It should contain only finite, non-missing, non-negative values, and the entries should sum to 1; if it is not, the vector is automatically normalized to sum to 1. By default, x0 is the vector with all equal values.

convtol.sqp
A small, non-negative number specifying the convergence tolerance for SQP algorithm; convergence is reached when the maximum dual residual in the Karush-Kuhn-Tucker optimality conditions is less than or equal to convtol.sqp. Smaller values will result in more stringent convergence criteria and more accurate solutions, at the expense of greater computation time.

convtol.activeset
Small, non-negative number specifying the convergence tolerance for the active-set step. Smaller values will result in higher quality search directions for the SQP algorithm but possibly a greater per-iteration computational cost.

zero.threshold
A small, non-negative number used to determine the "active set"; that is, it determines which entries of the solution are exactly zero. Any entries that are less than or equal to zero.threshold are considered to be exactly zero. Larger values of zero.threshold may lead to speedups for matrices with many columns, at the (slight) risk of prematurely zeroing some co-ordinates.

eps
A small, non-negative number added to the terms inside the logarithms to sidestep computing logarithms of zero. This prevents numerical problems at the cost of introducing a small inaccuracy in the solution. Increasing this number may lead to faster convergence but possibly a less accurate solution.

delta
A small, non-negative number added to the diagonal of the Hessian to improve numerical stability (and possibly the speed) when computing the search directions in the active-set step.

maxiter.sqp
Maximum number of SQP iterations to run before reporting a convergence failure; that is, the maximum number of quadratic subproblems that will be solved by the active-set method.

maxiter.activeset
Maximum number of active-set iterations taken to solve each of the quadratic subproblems.

verbose
If verbose = TRUE, the algorithm's progress and a summary of the optimization settings are printed to the console. The algorithm's progress is displayed in a table with one row per SQP (outer loop) iteration, and with the following columns: "iter", the (outer loop) SQP iteration; "objective", the value of the (unmodified) objective function at the current estimate of the solution; "max.diff", the maximum difference in the estimates between two successive iterations; "max(rdual)", the maximum "dual residual" in the Karush-Kuhn-Tucker (KKT) conditions, which is used to monitor convergence (see convtol.sqp); "nnz", the number of non-zero co-ordinates in the current estimate, as determined by zero.threshold, "nqp", the number of (inner loop) active-set iterations taken to solve the quadratic subproblem; "nls", the number of iterations in the backtracking line search.

x
The point at which the objective is evaluated in mixobjective; see argument x0 for details.

...
Additional arguments passed to KWDual.

Value

mixobjective returns the value of the (unmodified) objective at x. mixKWDual returns a list object with the following list elements:

x

The estimated solution to the convex optimization problem.

value

The value of the (unmodified) objective function at x.

status

The return status from MOSEK.

For more information on this output, see KWDual and mosek. mixSQP returns a list object with the following list elements:
x

The estimated solution to the convex optimization problem.

value

The value of the (unmodified) objective function at x.

status

A character string giving the status of the algorithm upon termination.

data

A data frame containing more detailed information about the algorithm's progress. The data frame has one row per SQP iteration. For an explanation of the columns, see the description of the verbose argument above. Missing values (NA's) in the last row indicate that these quantities were not computed because convergence was reached before computing them. Also note that the association of these quantities is slightly different than the console output when verbose = TRUE as the console output shows some quantities that were computed after the convergence check in the previous iteration. The last entries of max.diff, nqp and nls may not have been # assigned if the SQP algorithm converged successfully (as indicated # by negative values), in which case we should more appropriately # assign them missing values (NA).

Details

Here is a mathematical description of the constrained, convex optimization problem solved by mixSQP and mixKWDual. Let \(L\) be a matrix with \(n\) rows and \(m\) rows containing only non-negative entries, and let \(w = (w_1, \ldots, w_n)\) be a matrix of non-negative "weights" that sum to 1. mixSQP computes the value of vector \(x = (x_1, \ldots, x_m)\) minimizing the following objective function, $$f(x) = -1/n \sum_{j=1}^n w_j log (\sum_{k=1}^m L_{jk} x_k),$$ subject to the constraint that \(x\) lie within the simplex; that is, all entries of \(x\) are non-negative, and the sum of these entries is equal to 1. The Expectation Maximization (EM) algorithm can be used to solve this optimization problem, but it is intolerably slow in many interesting cases.

mixSQP is implemented using the Armadillo C++ linear algebra library, which can automatically take advantage of multithreaded matrix computations to speed up mixSQP for large L matrices, but only when R has been configured with a multithreaded BLAS/LAPACK library (e.g., OpenBLAS).

References

Y. Kim, P. Carbonetto, M. Stephens and M. Anitescu (2018). A fast algorithm for maximum likelihood estimation of mixture proportions using sequential quadratic programming. arXiv:1806.01412 https://arxiv.org/abs/1806.01412.

See also

KWDual

Examples

set.seed(1)
n  <- 1e5
m  <- 10
w  <- rep(1,n)/n
L  <- simulatemixdata(n,m)$L
out.mixsqp <- mixSQP(L,w)

#> Running mix-SQP 0.1-36 on 100000 x 10 matrix
#> convergence tol. (SQP):  1.0e-08
#> conv. tol. (active-set): 1.0e-10
#> max. iter (SQP):         1000
#> max. iter (active-set):  100
#> zero threshold:          1.0e-06
#> iter    objective max.diff max(rdual) nnz nqp nls
#>    1 +7.72664e-01       NA +2.756e-01  10  NA  NA
#>    2 +6.53047e-01 3.38e-01 +4.747e+00   5  13   1
#>    3 +6.04839e-01 4.67e-02 +2.163e+00   5   8   1
#>    4 +5.99876e-01 8.23e-03 +9.175e-01   5   2   1
#>    5 +5.99756e-01 1.59e-03 +2.977e-01   5   2   1
#>    6 +5.99697e-01 4.94e-03 +1.651e-01   5   2   1
#>    7 +5.99516e-01 9.17e-03 +7.180e-02   5   2   1
#>    8 +5.99192e-01 8.85e-03 +1.964e-02   5   2   1
#>    9 +5.98988e-01 3.64e-03 +1.972e-03   5   2   1
#>   10 +5.98959e-01 4.00e-04 +2.234e-05   5   2   1
#>   11 +5.98959e-01 4.45e-06 +2.888e-09   5   2   1
#> Convergence criteria met---optimal solution found.

out.kwdual <- mixKWDual(L,w)
print(mixobjective(L,out.mixsqp$x,w),digits = 16)

#> [1] 0.5989590363728408

print(mixobjective(L,out.kwdual$x,w),digits = 16)

#> [1] 0.5989590363746571