mixSQP

mixSQP solves a convex optimization problem in https://arxiv.org/abs/1806.01412. It implements a sequential quadratic programming with active-set subproblem solvers. For gigantic data, use low-rank approximation to speed up the computation.

When L is a (n) by (m) matrix of nonnegative entries, mixSQP maximizes the following objective function $$f(x) = \sum_j w_j log (\sum_k L_jk x_k)$$ subject to the (unit) probability simplex constraint $$\sum_k x_k = 1, x_k \ge 0$$ Without loss of generality \(\sum_j w_j = 1\) is required.

mixSQP(L, w = rep(1, nrow(L)), x0 = rep(1, ncol(L)), convtol = 1e-08,
  sparsetol = 1e-08, eps = .Machine$double.eps, maxitersqp = 1000,
  maxiteractiveset = 100, verbose = TRUE)

Arguments

L
Matrix specifying the optimization problem to be solved. It should be a numeric matrix with positive entries, and, ideally, double-precision.

w
A numeric vector, with one entry for each row of L, specifying the "weights" associated with the rows of L. All weights must be positive. It is assumed the weights sum to 1; if not, they will automatically be normalized to sum to 1. By default, all rows of L are assigned the same weight.

x0
An initial estimate of the solution to the optimization problem. It should contain only non-negative values, and the entries should sum to 1; if not, the vector is automatically normalized to sum to 1. By default, x0 is the vector with all equal entries.

convtol
A convergence tolerance used for algorithm's convergence criterion.

sparsetol
A tolerance used for determining active indices.

eps
A small constant to safeguard from a numerical issue (default 1e-6).

maxitersqp
Maximum number of SQP iterations; that is, the maximum number of quadratic subproblems that will be solved by the active-set method.

maxiteractiveset
Maximum number of active-set iterations taken to solve each of the quadratic subproblems.

verbose
If verbose = TRUE, print progress of algorithm to console.

Value

Returns a solution x (in the current version). Also returns the algorithm convergence status.

Examples

set.seed(1)
n  <- 1e5
m  <- 20
w  <- rep(1,n)/n
L  <- simulatemixdata(n,m)$L
x0 <- rep(1,m)/m # initialization
out <- mixSQP(L,w,x0)

#> Running mix-SQP 0.1-8 on 100000 x 20 matrix
#> convergence tolerance: 1.0e-08
#> zero threshold:        1.0e-08
#> max. SQP iter:         1000
#> max. active-set iter:  100
#> iter  objective   -min(g+1) #nnz #nqp #nls
#>    1  7.78710e-01 +2.48e-01   20 

#> Error in mixSQP_rcpp(L, w, x0, convtol, sparsetol, eps, maxitersqp, maxiteractiveset,     verbose): inv_sympd(): matrix is singular or not positive definite